DjangoApp/requirements.txt
django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
celery==5.2.1
